requirements.txt
# Mutual Fund FAQ Assistant - Python Dependencies
# Phase 1: Core dependencies for all phases

# ============================================
# Web Scraping (Phase 4.0)
# ============================================
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Optional: Playwright for JS-rendered pages
# playwright>=1.40.0

# ============================================
# Vector Store & Embeddings (Phase 4.2, 4.3, 5)
# ============================================
chromadb>=0.4.18
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0

# ============================================
# LLM Integration (Phase 6)
# ============================================
groq>=0.4.0
openai>=1.6.0

# ============================================
# API Framework (Phase 9)
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ============================================
# Data Processing (Phase 4.1, 4.2)
# ============================================
pydantic>=2.5.0
pyyaml>=6.0.1
python-dateutil>=2.8.0
tqdm>=4.66.0

# ============================================
# Thread Storage (Phase 8)
# ============================================
# SQLite is built-in, but for production Postgres:
# psycopg2-binary>=2.9.0
# asyncpg>=0.29.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6

# ============================================
# Testing & Quality
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# ============================================
# Type Stubs
# ============================================
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
types-pyyaml>=6.0.1
